gdb: improve equal check used by the pretty-printer [] support
When performing a BINOP_SUBSCRIPT via the pretty printers, if the
pretty printer represents a map, then the index field could be any
type.

This commit extends GDB so that we can use any compiled in equality
operators, or any available xmethods when comparing the index as
returned by the pretty-printer to the user supplied index, this allows
for better handling of C++ string based maps.

gdb/ChangeLog:

	* python/py-prettyprint.c: Add 'expop.h' include.
	(gdbpy_val_pretty_printer_find_child): Call eval_op_equal instead
	of value_equal.

gdb/testsuite/ChangeLog:

	* gdb.python/py-prettyprint-subscript.cc (class string_type): New
	class.
	(dump_bool): New global.
	(obj_str_int): New global.
	(main): Fill new global.
	* gdb.python/py-prettyprint-subscript.exp (run_tests): Add new
	tests.
	* gdb.python/py-prettyprint-subscript.py (class pp_string_type):
	New class.
	(class pretty_printer_finder): Add lookup for string to int type.
	(generic_array_subscript::worker::__call__): Add comment.
	(generic_map_subscript::worker::__call__): Add comment.
	(class string_type_xmethod): New class.