gdb/testsuite: resolve duplicate test name in gdb.cp/cplusfuncs.exp
While resolving duplicate test names I spotted that a test in
gdb.cp/cplusfuncs.exp included an unescaped '[]'.  In TCL square
brackets enclose expressions to evaluate, and so in this case, where
there is no enclosed expression, this just evaluates to the empty
string.

This clearly was not what the test intended, so in this commit I have
escaped the square brackets.  This has extended the test coverage.

gdb/testsuite/ChangeLog:

	* gdb.cp/cplusfuncs.exp (test_paddr_operator_functions): Escape
	square brackets in test.