[WIP] test for invalid debug ranges
This is a test for an issue I ran into on an older version of GDB, but
which was fixed in commit 520596f2eb7430dca76d887dc6c39698654af728.

That commit already has a test, so adding this test feels a little
redundant.

For now I'll just hold on to this test in case I need it again, but
I'll probably just delete this eventually.