gdb: avoid reusing type name as variable name in a few places
In a later commit I plan to make some changes to filename_seen_cache.
For some reason these changes cause GCC to start throwing an error
when confronted with:

  struct filename_seen_cache *filename_seen_cache;

I could investigate this more, but simply renaming the variable to NOT
have the same name as the type is an easy fix, and in this case it's
pretty self contained, so that's what I propose to do.

There should be no user visible changes after this commit.

gdb/ChangeLog:

	* symtab.c (struct output_source_filename_data): Rename
	filename_seen_cache field to filename_cache.
	(output_source_filename_data::output): Update for change in member
	variable name.
	(info_sources_command): Likewise.
	(struct add_partial_filename_data): Rename filename_seen_cache
	field to filename_cache.
	(add_partial_filename_data::operator()): Update for change in
	member variable name.
	(make_source_files_completion_list): Likewise.