gdb: avoid sending unnecessary wildcard vCont actions
Looking at remote_target::commit_resumed the algorithm for sending
wildcard actions starts by assuming that we can send a wildcard
action, and then looks for pending events, or resume requests, that
prevent wildcard actions (either global, or inferior specific).

This is fine, and there's no real problems with this, but it does lead
to one slight weirdness, if we are in non-stop mode, and
displaced-stepping is off, then, when we do a step, all threads should
be stopped, and we should step only the one thread that we wish to
have move on.  However, in the situation where there the inferior only
has a single thread then we actually end up sending a packet like
this:

  [remote] Sending packet: $vCont;s:pPID.TID;c#XX

Notice the trailing ';c'.

Now, there's nothing technically wrong here, we are correctly asking
the one and only thread PID.TID to perform a step, there are no other
threads, so telling everyone else to continue is harmless.

However, this annoys me.  It annoys me because we know we want only
the one, and only one, thread PID.TID to step, why send over redundant
information!

This commit fixes this tiny blip by keeping a counter of the number of
threads that could be resumed in each inferior.  We count up as we see
threads that need to be resumed, and then count down when non-wildcard
actions are sent (e.g. the s:pPID.TID in our example above).  Then,
when we are considering sending wildcard actions, we only send one if
there are any threads that will be impacted by the wildcard action.

Now, in the above situation we send out

  [remote] Sending packet: $vCont;s:pPID.TID#XX

There should be no change in behaviour when considering gdb/gdbserver
working together after this commit.

gdb/ChangeLog:

	* remote.c (struct remote_inferior): Add
	pending_vcont_resume_thread_count field.
	(remote_target::commit_resumed): Add use of
	pending_vcont_resume_thread_count field, only send wildcard action
	if there are threads the wildcard will cover.