gdb/python: move PyLong_From* calls into py-utils.c
We already have two helper functions in py-utils.c:

  gdb_py_object_from_longest (LONGEST l)
  gdb_py_object_from_ulongest (ULONGEST l)

these wrap around calls to either PyLong_FromLongLong,
PyLong_FromLong, or PyInt_From_Long (if Python 2 is being used).

There is one place in gdb/python/* where a call to PyLong_FromLong was
added outside of the above utility functions, this was done in the
recent commit:

  commit 55789354fcbaf879f3ca8475b647b2747dec486e
  Date:   Fri May 14 11:56:31 2021 +0200

      gdb/python: add a 'connection_num' attribute to Inferior objects

In this commit I propose that we move this new call to PyLong_FromLong
into a new utility function gdb_py_object_from_int.  The biggest win I
see here is that we have a consistent interface for converting number
like things into Python objects, however, I have made it such that on
Python 2 we will call PyInt_FromLong, so there is a slight change in
behaviour here, which seems to be inline with how we handle the
LONGEST types.

Beyond the PyLong/PyInt change described above there should be no user
visible changes from this commit.

gdb/ChangeLog:

	* python/py-inferior.c (infpy_get_connection_num): Call new
	function gdb_py_object_from_int.
	* python/py-utils.c (gdb_py_object_from_int): New function.
	* python/python-internal.h (gdb_py_object_from_int): Declare.