Make inferior/GDB share terminal in run+detach testcases
A following patch will make GDB put spawned inferiors in their own
terminal/session.  A consequence of that is that if you do "run", and
then "detach", GDB closes the terminal, and the inferior gets a
SIGHUP, after being detached, which likely means it'll die.  (The
referenced patch will make GDB detect the scenario and ask for
confirmation before detaching.)

To avoid that, testcases that explicitly want to do run/start + detach
need to tell GDB to not create a terminal for the inferior.  That's
what this patch does.

gdb/testsuite/ChangeLog:
yyyy-mm-dd  Pedro Alves  <pedro@palves.net>

	* gdb.base/attach-wait-input.exp: Issue "tty /dev/tty" before
	starting program.
	* gdb.base/auto-connect-native-target.exp: Likewise.
	* gdb/testsuite/gdb.base/detach.exp: Issue "tty /dev/tty" before
	starting program.  Use with_test_prefix.
	* gdb/testsuite/gdb.base/dprintf-detach.exp: Issue "tty /dev/tty"
	before starting program.
	* gdb/testsuite/gdb.base/jit-elf.exp: Likewise.
	* gdb/testsuite/gdb.base/multi-forks.exp: Likewise.
	* gdb/testsuite/gdb.base/watchpoint-hw-attach.exp: Likewise.
	* gdb/testsuite/gdb.mi/mi-detach.exp: Likewise.
	* gdb/testsuite/gdb.multi/watchpoint-multi-exit.exp: Likewise.
	* gdb.threads/process-dies-while-detaching.exp: Likewise.
	* gdb.threads/threadapply.exp: Likewise.

Change-Id: Ic62bca178295763fb9c47657ee459fe715f7865e