gdb/remote: error if a stop arrives for a non-resumed thread
When looking at bug gdb/26819 I spotted that due to the configuration
of OpenOCD, OpenOCD was sending GDB a stop reply for a thread that GDB
was not expecting to be executing.

If GDB was doing a non-displaced step-over, and so, had only sent a
single step request to one specific thread, GDB would, assert that the
thread that stopped should be in the process of completing a
step-over.  This would result in this assertion:

  infrun.c:5672: internal-error: int finish_step_over(execution_control_state*): Assertion `ecs->event_thread->control.trap_expected' failed.

This commit is an attempt to have GDB spot when the incoming stop
packet from the target is invalid and throw an error before we hit the
above assert.

So, in this commit, when processing the stop in the remote_target
code, if the stop is for a thread that is not resumed, throw an error.

There's no test for this, as it requires that a target send back an
invalid stop packet, one that includes the wrong thread-id, I don't
currently have any good ideas for how I would trigger such a remote
packet.

When looking at the original example from the bug report (which uses
the RISC-V spike simulator and OpenOCD), when the invalid packet
arrives the GDB session now looks like this:

  (gdb) continue
  stop received from remote for thread that should not be executing
  (gdb) info threads
    Id   Target Id                                       Frame
    1    Thread 1 (Name: riscv.cpu0, state: single-step) (running)
  * 2    Thread 2 (Name: riscv.cpu1, state: single-step) (running)

Unfortunately, at this point the session is effectively dead as the
threads are still in the running state, but GDB is in all-stop mode,
so isn't really expecting to be in this state.

Given that, should we get to this point, we know that the remote is
not following the rules, I'm not hugely worried.  Even if we could put
GDB back into a usable state, we couldn't reasonable expect to
continue working with the remote when we know it's broken.

gdb/ChangeLog:

	PR gdb/26819
	* remote.c (remote_target::process_stop_reply): Catch the case
	where a stop reply arrives for a thread that was not resumed.