gdb: support stepping over a clone syscall with displaced stepping off
Not to be confused with bug gdb/19675, which is stepping over a clone
syscall when using displaced stepping, this commit adds support for
stepping over a clone syscall when using in-place stepping.

Currently, when a new thread is created through a clone syscall, GDB
will set the new thread running.  With 'continue' this makes sense:

 - all-stop mode, user issues 'continue', all threads are set running,
   a newly created thread should also be set running.

 - non-stop mode, user issues 'continue', other pre-existing threads
   are not effected, but as the new thread is (sort-of) a child of the
   thread the user asked to run, it makes sense (to me) that the new
   threads should be created in the running state.

Similarly, if we are stopped at the clone syscall, and there's no
software breakpoint at this address, then the current behaviour is
fine I think:

 - all-stop mode, user issues 'stepi', stepping will be done in
   place (as there's no breakpoint to step over).  While stepping the
   thread of interest all the other threads will be allowed to
   continue.  A newly created thread will be set running, and then
   stopped once the thread of interest has completed its step.

 - non-stop mode, user issues 'stepi', stepping will be done in
   place (as there's no breakpoint to step over).  Other threads might
   be running or stopped, but as with the continue case above, the new
   thread will be created running.  The only possible issue here is
   that the new thread will be left running after the initial thread
   has completed its stepi.  The user would need to manually select
   the thread and interrupt it, this might not be what the user
   expects.  However, this is not something I'm trying to change in
   this commit.

The problem then is what happens when we try to step over a clone
syscall if there is a software breakpoint at the syscall address.

We can ignore the displaced stepping case, this is covered by bug
gdb/19675.  For this commit we are only interested in the
non-displaced stepping case:

 - For both all-stop and non-stop modes:
   + user issues 'stepi',
   + [non-stop mode only] GDB stops all threads.  In all-stop mode all
     threads are already stopped.
   + GDB removes s/w breakpoint at syscall address,
   + GDB single steps just the thread of interest, all other threads
     are left stopped,
   + New thread is created running,
   + Initial thread completes its step,
   + [non-stop mode only] GDB resumes all threads that it previously
     stopped.

There are two problems here:

  1. The new thread might pass through the same code that the initial
     thread is in (i.e. the clone syscall code), in which case it will
     fail to hit the breakpoint in clone as this was removed so the
     first thread can single step,

  2. The new thread might trigger some other stop event before the
     initial thread reports its step complete.  If this happens we end
     up triggering an assertion as GDB assumes that only the thread
     being stepped should stop.  The assert looks like this:

     infrun.c:5899: internal-error: int finish_step_over(execution_control_state*): Assertion `ecs->event_thread->control.trap_expected' failed.

What is needed is for GDB to have more control over whether the new
thread is created running or not.  Issue #1 above requires that the
new thread not be allowed to run until the s/w breakpoint has been
reinserted, the only way to guarantee this is if the new thread is
held in a stopped state until the single step has completed.

When looking for solutions to this problem I considered how GDB
handles fork/vfork as these seemed like they should have some of the
same issues.  The main difference I see between fork/vfork and clone
is that the thread created events are not, by default, reported back
to core GDB, instead the thread created event is handled automatically
in the target code and the thread is immediately set running.

However, GDB does have a mechanism to request that it be told about
new threads, this is target_ops::thread_events.  This on/off flag,
when on, indicates that the targets should notify GDB about new
threads (and also threads exiting).  The new threads are still created
in a "running" state, that is, the infrun core of GDB sees the thread
as running, but the target will not actually start the thread
executing, this is perfect for what we need.

My proposal is that, before GDB steps a thread (when it is stepping
over a software breakpoint), we should first enable thread_events,
then, once the step has completed, disable thread_events.

What this means is that the new thread will be created in the infrun
running state, but not set executing.  Once the step has completed
then:

  - In all-stop mode
    + the stepping thread will stop,
    + all other threads (except the new one) are already stopped,
    + the new thread is not executing, but is marked running, however
      GDB already calls stop_all_threads, which moves the thread into
      the stopped state,
    + we're done.

  - In non-stop mode
    + the stepping thread will stop,
    + all other threads (except the new one) are already stopped,
    + the new thread is not executing, but is marked running,
    + GDB calls restart_threads, the new thread will be set executing,
    + we're done.

This change worked fine for the native Linux target, but ran into
issues with the remote target.

The problem is that when enabling thread events for the remote target,
the event isn't actively reported to GDB, that is, if thread-events is
on, and a new thread is created, then gdbserver does hold the thread
in the stopped state, but doesn't report the stop to GDB.

This behaviour works fine for the current use of the thread-events
mechanism, which is stop_all_threads (infrun.c), in this case (for
remote targets) we're going to end up pulling the stop packets using
the vStopped mechanism anyway, so having gdbserver actively push the
stop to GDB is not important.

However, this causes problems for me, we enable thread-events and step
the target.  The new thread is created but held in a stopped
state (with a pending thread-created event), when GDB resumes the
target the thread-created stop is immediately reported back to GDB.

My proposal in this commit is to have gdbserver discard pending
thread-created events if GDB is not interested in being told about
these events.  This filtering is done earlier, at the point where we
pull pending stop events from the threads, prior to resuming any
threads.  The consequence of this is that if GDB has said it is no
longer interested in thread creation events, any pending events are
discarded, and the thread is resumed.

I've tested this on GNU/Linux using the native target, and with the
native-gdbserver and native-extended-gdbserver targets.

gdb/ChangeLog:

	PR gdb/27830
	* infrun.c (displaced_step_finish): Disable thread events.
	(do_target_resume): Enable thread events when stepping.
	(finish_step_over): Disable thread events.

gdb/testsuite/ChangeLog:

	PR gdb/27830
	* gdb.threads/stepi-over-clone.c: New file.
	* gdb.threads/stepi-over-clone.exp: New file.

gdbserver/ChangeLog:

	PR gdb/27830
	* linux-low.cc
	(linux_process_target::thread_still_has_status_pending): Add
	support for filtering out unwanted thread created events.