gdb: support stepping over a clone syscall with displaced stepping off
Not to be confused with bug gdb/19675, which is stepping over a clone
syscall when using displaced stepping, this bug is for stepping over a
clone syscall when using in-place stepping.

Currently, when a new thread is created through a clone syscall, GDB
will set the new thread running.  With 'continue' this makes sense:

 - all-stop mode, user issues 'continue', all threads are set running,
   a newly created thread should also be set running.

 - non-stop mode, user issues 'continue', it makes sense that a new
   thread is set running.

Similarly, if we are stopped at the clone syscall, and there's no
software breakpoint at this syscall, then the current behaviour is
fine I think:

 - all-stop mode, user issues 'stepi', stepping will be done in
   place (as there's no breakpoint to step over).  While stepping the
   thread of interest all the other threads will be allowed to
   continue.  A newly created thread will be set running, and then
   stopped once the thread of interest has completed its step.

 - non-stop mode, user issues 'stepi', stepping will be done in
   place (as there's no breakpoint to step over).  Other threads might
   be running or stopped, but I don't think it is incorrect for the
   new thread to be created running, the user would then need to
   switch to the new thread and manually stop it if required.

The problem then is what happens when we try to step over a clone
syscall if there is a software breakpoint at the syscall address.  We
can ignore the displaced stepping case, this is covered by bug
gdb/19675, for this commit we are only interested in the non-displaced
stepping case:

 - For both all-stop and non-stop modes:
   + user issues 'stepi',
   + GDB removes s/w breakpoint at syscall address,
   + [non-stop mode only] GDB stops all threads,
   + GDB single steps just the thread of interest, all other threads
     are left stopped,
   + New thread is created running,
   + Initial thread completes its step,
   + [non-stop mode only] GDB resumes all threads that it previously
     stopped.

There are two problems here:

  1. The new thread might pass through the same code that the initial
     thread is in (i.e. the clone syscall code), in which case it will
     fail to hit the breakpoint in clone as this was removed so the
     first thread can single step,

  2. The new thread might trigger some other stop event before the
     initial thread reports its step complete.  If this happens we end
     up with this assert triggering:

     infrun.c:5899: internal-error: int finish_step_over(execution_control_state*): Assertion `ecs->event_thread->control.trap_expected' failed.

What we want if for GDB to be able to apply more logic when deciding
if the new thread should be set executing or not.  If we compare clone
to fork/vfork, we can see that the difference is that clone events are
not routinely reported to the GDB core, instead they are mostly
handled in the target code, the new thread is spotted, the target adds
the new thread to GDB's thread database, then the thread is set
executing.  The infrun core of GDB never gets a chance to decide if
the new thread should be executing or not.

However, GDB does have a mechanism to request that it be told about
new threads, this is target_ops::thread_events.  This on/off flag,
when on, indicates that the targets should notify GDB about new
threads (and also threads exiting).  The new threads are still created
in a "running" state, that is, the infrun core of GDB sees the thread
as running, but the target will not actually start the thread
executing, this is perfect for what I want.

My proposal is that, before GDB steps a thread (when it is stepping
over a software breakpoint), we should first enable thread_events,
then, once the step has completed, disable thread_events.

What this means is that the new thread will be created in the infrun
running state, but not set executing.  Once the step has completed
then:

  - In all-stop mode
    + the stepping thread will stop,
    + all other threads (except the new one) are already stopped,
    + the new thread is not executing, but is marked running, however
      GDB already calls stop_all_threads, which moves the thread into
      the stopped state,
    + we're done.

  - In non-stop mode
    + the stepping thread will stop,
    + all other threads (except the new one) are already stopped,
    + the new thread is not executing, but is marked running,
    + GDB calls restart_threads, the new thread will be set executing,
    + we're done.

I've tested this on GNU/Linux using the native target, and with the
native-gdbserver and native-extended-gdbserver targets.

gdb/ChangeLog:

	PR gdb/27830
	* infrun.c (displaced_step_finish): Disable thread events.
	(do_target_resume): Enable thread events when stepping.
	(finish_step_over): Disable thread events.

gdb/testsuite/ChangeLog:

	PR gdb/27830
	* gdb.threads/stepi-over-clone.c: New file.
	* gdb.threads/stepi-over-clone.exp: New file.