gdb: non-displaced step over clone, remote target, no QThreadEvents
This commit compliments the previous commit.  While the previous
commit aims to offer a solution for non-displaced stepping over a
clone syscall, the solution depends on a target supporting the
QThreadEvents packet.

As this packet is optional we should consider what would happen if a
target does not support this packet.

The short answer is that, without QThreadEvents support, GDB can't
correctly support non-displaced stepping of a clone syscall, all of
the bugs discussed in the previous commit can still be hit.

However, when GDB is in non-stop mode, and we do a non-displaced step
over a clone syscall instruction, for a remote target that does not
support QThreadEvents, then we run into an error, like this:

  (gdb) stepi
  [New Thread 2914644.2914648]
  PC register is not available

The 'PC register is not available' is an error which is caused by GDB
attempting to read the $pc from a running thread.

The problem here is that when the remote target sees the new thread,
the thread is created in the running state, this is good because we
know that the remote will have set the new thread running already.

So, the thread_info::executing field will be true (indicating that the
target believes the thread is executing), but the thread_info::resumed
field will be false as infrun did not resume the thread.

Now in non-stop mode, when performing a non-displaced step, all
threads are stopped, once the step is complete GDB calls
infrun.c:restart_threads to restart all of the threads that GDB
previously stopped.  The restart_threads function loops over all
threads, and looks that the thread state, if the thread needs
restarting then this is done.

However, the restart_threads function does not consider the case where
a thread might be executing, but not resumed, and so GDB tries to
restart the already executing thread, this requires a read of $pc,
which then fails.

In this commit I propose to add a new case to restart_threads, this
will spot the case where a new thread is already executing, but not
resumed.  This thread will then be moved in to the resumed state with
no further action required.  That is we acknowledge that the thread is
already executing, and just infrun just updates its internal state so
it thinks that it asked the thread to resume.

I updated an existing test to cover this case and tested using native
GNU/Linux target, native-gdbserver, and native-extended-gdbserver.

gdb/ChangeLog:

	PR gdb/27830
	* infrun.c (restart_threads): Threads that are already executing
	in the target are already resumed.

gdb/testsuite/ChangeLog:

	PR gdb/27830
	* gdb.threads/stepi-over-clone.exp: Add additional tests.