gdb: print backtrace on fatal SIGSEGV
This commit adds a new maintenance feature, the ability to print
a (limited) backtrace if GDB dies due to a fatal signal.

The backtrace is produced using the backtrace and backtrace_symbols_fd
functions which are declared in the execinfo.h header.  A configure
check has been added to check for these features, if they are not
available then the new code is not compiled into GDB.

The motivation for this new feature is to aid in debugging GDB in
situations where GDB has crashed at a users site, but the user is
reluctant to share core files, possibly due to concerns about what
might be in the memory image within the core file.  Such a user might
be happy to share a simple backtrace that was written to stderr.

The production of the backtrace is off by default, and can switched on
using the new commands:

  maintenance set backtrace-on-fatal-signal on|off
  maintenance show backtrace-on-fatal-signal

Right now, I have hooked this feature in to GDB's existing handling of
SIGSEGV only, but this could be extended to more signals in a later
commit.

One additional change I have made in this commit is that, when we
decide that GDB should terminate due to the fatal signal we now
set the signal disposition of the signal in question back to SIG_DFL,
unblock the signal, and then re-raise the signal.  This ensures that
GDB should terminate due to the specific bad signal that arrived.

So, currently, this is only effecting our handling of SIGSEGV.
Previously, if GDB hit a SEGV then we would terminate GDB with a
SIGABRT.  After this commit we will terminate GDB with a SIGSEGV.
This feels like an improvement to me, we should still get a core dump,
but in many shells, the user will see a more specific message once GDB
exits, in bash for example "Segmentation fault" rather than "Aborted".

gdb/ChangeLog:

	* NEWS: Mention new commands.
	* config.in: Regenerate.
	* configure: Regenerate.
	* configure.ac: Add header check for execinfo.h.  Add a check for
	the backtrace and backtrace_symbols_fd functions.
	* event-top.c: Add execinfo.h include.
	(bt_on_fatal_signal): New static global.
	(show_bt_on_fatal_signal): New function.
	(set_bt_on_fatal_signal): New function.
	(unblock_signal): New function.
	(handle_fatal_signal): New function.
	(handle_sigsegv): Call handle_fatal_signal instead of abort.
	(async_sigtstp_handler): Make use of unblock_signal.
	(_initialize_event_top): Register new maintenance command.
	* ui-file.h (ui_file) <fd>: New virtual member function.
	(stdio_file) <fd>: New member function.

gdb/doc/ChangeLog:

	* gdb.texinfo (Maintenance Commands): Document new commands.

gdb/testsuite/ChangeLog:

	* gdb.base/bt-on-fatal-signal.c: New file.
	* gdb.base/bt-on-fatal-signal.exp: New file.